% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/step0.5_capitalise_fast_parameter.R
\name{B_fastparameteroptions}
\alias{B_fastparameteroptions}
\title{B_fastparameteroptions
Function author: TomW
Last edit date: 14 Jan 24
updates parameteroptions to 'fast' options, for debugging}
\usage{
B_fastparameteroptions(parameterOptions)
}
\arguments{
\item{parameterOptions}{parameterOptions is loaded as part of the package}
}
\value{
dataframe, with fast parameter options for code testing
}
\description{
B_fastparameteroptions
Function author: TomW
Last edit date: 14 Jan 24
updates parameteroptions to 'fast' options, for debugging
}
\examples{
fastparameteroptions(parameterOptions)
}
