% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadExternal.R
\name{downloadExternal}
\alias{downloadExternal}
\title{Download the cutadapt.exe file and SILVA 138.1 taxa
reference database from the SAMS THREDDS server, or
from source.}
\usage{
downloadExternal()
}
\description{
Download the cutadapt.exe file and SILVA 138.1 taxa
reference database from the SAMS THREDDS server, or
from source.
}
\section{Example usage}{
 eDNA2IQI::downloadExternal()
}

