% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-scatterplot.R
\name{draw_nmds_plot}
\alias{draw_nmds_plot}
\title{Draw scatter plot}
\usage{
draw_nmds_plot(nmds_test, folder)
}
\arguments{
\item{nmds_test}{output from internal nmds_bray() function.}

\item{folder}{active folder}
}
\description{
Draw NMDS scatter plot of observed data compared to training data,
highlighting outliers.
}
\keyword{internal}
