% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawBarplot.R
\name{drawBarplot}
\alias{drawBarplot}
\title{drawBarplot}
\usage{
drawBarplot(S16_readsB, folder)
}
\arguments{
\item{S16_readsB}{annotated reads for plotting from}

\item{folder}{active folder}
}
\description{
Takes S16_readsB from step2, and step3 and saves unrarefied and rarefied taxa barplot respectively
}
\keyword{internal}
