% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds-bray.R
\name{nmds_bray}
\alias{nmds_bray}
\title{Compare Non-parametric Multidimensional Scaling (NMDS) Centroid Distances}
\usage{
nmds_bray(
  data,
  training_data,
  trymax = 20,
  probs = 0.95,
  data_output_only = TRUE
)
}
\arguments{
\item{data}{Observed data}

\item{training_data}{Data from training set.}

\item{trymax}{Maximum number of random starts in search of stable solution}

\item{probs}{Set percentile distance}

\item{data_output_only}{Return observed data only, if FALSE, training data is
also returned}
}
\value{
Dataframe of observed samples.
}
\description{
The relationship between IQI and bacteria within the random forest model is
based on data collected to train the model. Sometimes if the data
observed is significantly different to the data in the training set, we
should have less confidence in the prediction.
}
\details{
To test the applicability of the data observed for use with the model, we
compare the distance to NMDS coordinates of each sample in the observed data
to the 95th percentile distance to the centroid of samples in the training
set. Samples are flagged in the observed data which are more than 95th
percentile distance to centroid compared to the training set.

Samples may not be applicable for the model due to environmental, sampling,
analysis or data processing anomalies. The cause should be investigated and
remedial steps taken where required.
}
\keyword{internal}
