% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1.01_quality_check.r
\name{step1.01_quality_check}
\alias{step1.01_quality_check}
\title{Check quality of raw data, output graphs}
\usage{
step1.01_quality_check(folder)
}
\arguments{
\item{folder}{A string. Location of raw data files.}
}
\value{
NULL output saved directly to folder.
}
\description{
Check quality of raw data, output graphs
}
