% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1.1_extract_identifier.r
\name{step1.1_extract_identifier}
\alias{step1.1_extract_identifier}
\title{Extract the initial line from fastq.gz files and report data frame with
unique instrument_run pairs and sample names.}
\usage{
step1.1_extract_identifier(folder)
}
\arguments{
\item{folder}{A string. Location where the raw fastq.gz files can be found.}
}
\value{
A data frame.
}
\description{
Extract the initial line from fastq.gz files and report data frame with
unique instrument_run pairs and sample names.
}
