% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1.2_remove_unspecified.r
\name{step1.2_remove_unspecified}
\alias{step1.2_remove_unspecified}
\title{Remove unspecified or ambiguous bases from the sequence reads.}
\usage{
step1.2_remove_unspecified(folder, raw_files)
}
\arguments{
\item{folder}{A string. Location of raw data files.}

\item{raw_files}{A string. List of files from same location.}
}
\value{
A list.
}
\description{
Remove unspecified or ambiguous bases from the sequence reads.
}
