% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/step1.3_trim_primers.r
\name{Step1.3_trim_primers}
\alias{Step1.3_trim_primers}
\title{Step1.2_trim_primers
Remove the primer sequences from the sequence reads}
\usage{
Step1.3_trim_primers(folder, fwd_no_unspec, rev_no_unspec, parameteroptions)
}
\arguments{
\item{folder}{A string. Location of raw data files.}

\item{fwd_no_unspec}{A character vector. Contains the file location of
forward reads of the input samples where Ns were removed.}

\item{rev_no_unspec}{A character vector. Contains the file location of
reverse reads of the input samples where Ns were removed.}

\item{parameteroptions}{A dataframe.}
}
\value{
A list.
}
\description{
Step1.2_trim_primers
Remove the primer sequences from the sequence reads
}
