% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/step1.4_filter_quality_chimera.r
\name{Step1.4_filter_quality_chimera}
\alias{Step1.4_filter_quality_chimera}
\title{Step1.4_filter_quality_chimera}
\usage{
Step1.4_filter_quality_chimera(
  folder,
  raw_files,
  fwd_trimmed,
  rev_trimmed,
  parameteroptions
)
}
\arguments{
\item{folder}{A string. Location of raw data files.}

\item{raw_files}{A string. List of files from same location.}

\item{fwd_trimmed}{A character vector. the file name of forward
reads of the input samples where primer sequences were removed.}

\item{rev_trimmed}{A character vector. Contains the file location of reverse
reads of the input samples where primer sequences were removed.}

\item{parameteroptions}{A dataframe.}
}
\value{
A data frame.
}
\description{
Step1.4_filter_quality_chimera
}
