% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.0_annotate_asvs.r
\name{step2.0_annotate_ASVs}
\alias{step2.0_annotate_ASVs}
\title{Title Step2.0_annotate_ASVs}
\usage{
step2.0_annotate_ASVs(
  folder,
  parameteroptions,
  collated_asv_batches,
  auto_download = FALSE
)
}
\arguments{
\item{folder}{A string. Location of raw fastq files}

\item{parameteroptions}{A dataframe,}

\item{collated_asv_batches}{result from Step1..., dataframe}

\item{auto_download}{Boolean to override console prompt to automatically
download external dependencies. This can be useful for automating or
testing this function.}
}
\value{
A dataframe, written to file, combing ASVs across multiple batches
}
\description{
Title Step2.0_annotate_ASVs
}
\section{Example usage}{
 myTAXA <- step2.0_annotate_ASVs(folder,parameteroptions,myASVs)
}

