% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.0_predict_iqi.r
\name{step3.0_predict_iqi}
\alias{step3.0_predict_iqi}
\title{Generates the predicted IQI values based on the optimized random forest
model.}
\usage{
step3.0_predict_iqi(folder, S16_reads)
}
\arguments{
\item{folder}{A string. Location of raw fastq data files.}

\item{S16_reads}{A data frame. Contains the bacterial taxa read counts which
should be used to predict the IQI values.}
}
\value{
A data frame. As inputted, with added 'predicted_IQI' column
}
\description{
Generates the predicted IQI values based on the optimized random forest
model.
}
\section{Example usage}{
 myPREDs <- step3.0_predict_iqi(folder = "C:/full/file/path/to/fastq/S16_reads", S16_reads = myTAXA)
}

