% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/edna2iqi-package.r
\docType{package}
\name{eDNA2IQI-package}
\alias{eDNA2IQI}
\alias{eDNA2IQI-package}
\title{eDNA2IQI: Predict IQI from eDNA}
\description{
Tidy and clean bacterial DNA sequence reads. Allocate taxa names to unique sequences, using a curated SILVA v138.1 database as a reference. Predict IQI using a trained random forest algorithm.
}
\author{
\strong{Maintainer}: Adam Wyness \email{adam.wyness@sams.ac.uk} (\href{https://orcid.org/0000-0003-1202-6358}{ORCID})

Authors:
\itemize{
  \item Stevie Brain (\href{https://orcid.org/0000-0001-8272-6996}{ORCID})
  \item Esmee Roosenmaallen (\href{https://orcid.org/0000-0003-0519-3721}{ORCID})
  \item Tom Wilding \email{tom.wilding@sams.ac.uk}
}

}
\keyword{internal}
