% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edna2iqi.r
\name{edna2iqi}
\alias{edna2iqi}
\title{Wrapper function to perform all steps of reading and denoising raw reads,
allocating taxa and predicting IQI from multiple random forests}
\usage{
edna2iqi(
  folder,
  parameteroptions,
  auto_download = FALSE,
  remove_outliers = TRUE
)
}
\arguments{
\item{folder}{A string. Location of raw fastq data files.}

\item{parameteroptions}{A dataframe. Values for all function parameters.}

\item{auto_download}{Boolean to override console prompt to automatically
download external dependencies. This can be useful for automating or
testing this function.}

\item{remove_outliers}{Remove outliers detected by \code{?eDNA2IQI:::nmds_test()}.}
}
\value{
dataframes containing denoised ASVs (myASVs),
allocated taxa (myTaxa) and
predicted IQI values (myPreds)
}
\description{
Wrapper function to perform all steps of reading and denoising raw reads,
allocating taxa and predicting IQI from multiple random forests
}
\section{Example usage}{

edna2iqi(folder = "C:/full/file/path/to/fastq/", parameteroptions = parameteroptions)
}

