% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edna2iqi.r
\name{edna2iqi}
\alias{edna2iqi}
\title{Wrapper function to perform all steps of Reading, denoising,
allocating taxa and predicting IQI}
\usage{
edna2iqi(folder, parameteroptions, auto_download = FALSE)
}
\arguments{
\item{folder}{A string. Location of raw fastq data files.}

\item{parameteroptions}{A dataframe. Values for all function parameters.}

\item{auto_download}{Boolean to override console prompt to automatically
download external dependencies. This can be useful for automating or
testing this function.}
}
\value{
dataframes containing denoised ASVs (myASVs),
allocated taxa (myTaxa) and
predicted IQI values (myPreds)
}
\description{
Wrapper function to perform all steps of Reading, denoising,
allocating taxa and predicting IQI
}
\section{Example usage}{

edna2iqi(folder = "C:/full/file/path/to/fastq/", parameteroptions = parameteroptions)
}

