% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/extract_parameteroptions.r
\name{ExtractParameters}
\alias{ExtractParameters}
\title{Title ExtractParameters
Extracts parameters from parameteroptions dataframe provided with eDNA2IQI
Values are stored as characters, this returns a string or
NA, NULL, TRUE/FALSE if "Char2Vect" included in the ...}
\usage{
ExtractParameters(parameteroptions, FunctionName, ParameterName, ...)
}
\arguments{
\item{parameteroptions}{the parameteroptions dataframe (tibble), care on capitalisation}

\item{FunctionName}{which function in dada2 etc (e.g. trim etc)}

\item{ParameterName}{the parameters within the function}

\item{...}{optional modifications (?necessary)}
}
\value{
a character, logical or numerical value
}
\description{
Title ExtractParameters
Extracts parameters from parameteroptions dataframe provided with eDNA2IQI
Values are stored as characters, this returns a string or
NA, NULL, TRUE/FALSE if "Char2Vect" included in the ...
}
\examples{
ExtractParameters(parameteroptions,"cutadapt","-m")
}
