% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParameters.r,
%   R/extract_parameteroptions.r
\name{extractParameters}
\alias{extractParameters}
\title{extractParameters}
\usage{
extractParameters(parameteroptions, FunctionName, ParameterName, ...)

extractParameters(parameteroptions, FunctionName, ParameterName, ...)
}
\arguments{
\item{parameteroptions}{the parameteroptions dataframe (tibble), care on capitalisation}

\item{FunctionName}{which function in dada2 etc (e.g. trim etc)}

\item{ParameterName}{the parameters within the function}

\item{...}{optional modifications (?necessary)}
}
\value{
a character, logical or numerical value

a character, logical or numerical value
}
\description{
Extracts parameters from parameteroptions dataframe provided with eDNA2IQI
Values are stored as characters, this returns a string or
NA, NULL, TRUE/FALSE if "Char2Vect" included in the ...
}
\section{Example usage}{
 extractParameters(parameteroptions,"cutadapt","-m", "Char2Vect")
}

\examples{
extractParameters(parameteroptions,"cutadapt","-m")
}
