% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step0.5_capitalise_fast_parameter.R
\name{fastparameteroptions}
\alias{fastparameteroptions}
\title{fastparameteroptions
Function author: TomW}
\usage{
fastparameteroptions(parameterOptions)
}
\arguments{
\item{parameterOptions}{parameterOptions is loaded as part of the package}
}
\value{
dataframe, with fast parameter options for code testing
}
\description{
updates parameteroptions to 'fast' options, for debugging
}
