% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/messagecolour.r
\name{messageColour}
\alias{messageColour}
\title{Custom colour messages and warnings}
\usage{
messageColour(text, version)
}
\arguments{
\item{text}{A string. Text that is to be outputted.}

\item{version}{A string. Specifies which colour setting to use.}
}
\description{
Custom colour messages and warnings
}
