% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{parameteroptions}
\alias{parameteroptions}
\title{Parameter options dataset}
\format{
\subsection{\code{parameteroptions}}{

A data frame with 52 rows and 4 columns:
\describe{
\item{Function}{Function name parameter belongs to}
\item{Parameter}{Parameter name}
\item{Value}{Parameter value}
\item{Description}{Description of parameter, may include default value}
}
}
}
\usage{
parameteroptions
}
\description{
A dataset listing all parameter options for denoising functions.
}
\details{
Dataset used in examples and testing to configure denoising steps in
the \code{edna2iqiRFtrainer} package.
}
\keyword{datasets}
