% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadExternal.R
\name{print_consent_box}
\alias{print_consent_box}
\title{Download the random forest models,  cutadapt.exe file and SILVA 138.1 taxa
reference database from the SAMS THREDDS server, or
from source.}
\usage{
print_consent_box()
}
\arguments{
\item{auto_download}{Boolean to override console prompt to automatically
download external dependencies. This can be useful for automating or
testing this function.}
}
\description{
Download the random forest models,  cutadapt.exe file and SILVA 138.1 taxa
reference database from the SAMS THREDDS server, or
from source.
}
\section{Example usage}{
 eDNA2IQI::downloadExternal()
}

