% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step0.5_capitalise_fast_parameter.R
\name{step0.5_capitaliseR1_R2}
\alias{step0.5_capitaliseR1_R2}
\title{step0.5_capitaliseR1_R2}
\usage{
step0.5_capitaliseR1_R2(folder)
}
\arguments{
\item{folder, }{string naming the location of the raw FASTQ files to be processed}
}
\value{
NULL, code changes the capitalisation on the input files
}
\description{
Reads in the input data, capitalises any "r1" and "r2" in the file metadata componenets of
an Illumina Miseq output filename
This addresses software and filesharing bugs that lose case sensitivity, resulting in the
conversion of filenames to lower case.
The read orientation, indicated by R1 or R2 is essential for the bioinformatic process, and making it the
correct case saves further unnessecarily flexible code.
}
\examples{
step0.5_capitaliseR1_R2(folder)
}
