% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/step0.5_capitalise_fast_parameter.R
\name{Step0.5_capitaliseR1_R2}
\alias{Step0.5_capitaliseR1_R2}
\title{A_CapitaliseR1_R2
Reads in the input data, capitalises any r1 and r2 in string L001_R1 etc
convenient function, to address MS Explorer converting files to lower case during transfer.}
\usage{
Step0.5_capitaliseR1_R2(fastqfolder)
}
\arguments{
\item{fastqfolder}{string naming the location of the files to be processed}
}
\value{
NULL, code changes the capitalisation on the input files
}
\description{
A_CapitaliseR1_R2
Reads in the input data, capitalises any r1 and r2 in string L001_R1 etc
convenient function, to address MS Explorer converting files to lower case during transfer.
}
\examples{
CapitaliseR1_R2(path_to_fastq_folder)
}
