% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/step1.01_quality_check.r
\name{Step1.01_quality_check}
\alias{Step1.01_quality_check}
\title{Check quality of raw data, output graphs}
\usage{
Step1.01_quality_check(folder)
}
\arguments{
\item{folder}{A string. Location of raw data files.}
}
\description{
Check quality of raw data, output graphs
}
