% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1.0_readfastq.R
\name{step1.0_readFastq}
\alias{step1.0_readFastq}
\title{Read raw fastq.gz files, clean sequence reads, allocate taxa at level
specified. Save dataframe of S16 reads (as .rda), save dataframe of taxa
allocated reads (as .csv).}
\usage{
step1.0_readFastq(folder, parameteroptions)
}
\arguments{
\item{folder}{A string. Location of raw fastq data files.}

\item{parameteroptions}{A dataframe. Values for all function parameters.}
}
\value{
A data frame. Table of reads per taxa, sorted by sample.
}
\description{
Read raw fastq.gz files, clean sequence reads, allocate taxa at level
specified. Save dataframe of S16 reads (as .rda), save dataframe of taxa
allocated reads (as .csv).
}
\section{Example usage}{

data <- step1.0_readFastq(folder = "C:/full/file/path/to/fastq/")
}

