% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/step1.2_remove_unspecified.r
\name{Step1.2_remove_unspecified}
\alias{Step1.2_remove_unspecified}
\title{Remove unspecified or ambiguous bases from the sequence reads.}
\usage{
Step1.2_remove_unspecified(folder, raw_files)
}
\arguments{
\item{folder}{A string. Location of raw data files.}

\item{raw_files}{A string. List of files from same location.}
}
\value{
A list.
called by readFastq_TW (under 'MAIN LOOP')
}
\description{
Remove unspecified or ambiguous bases from the sequence reads.
}
