% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1.4_filter_quality_chimera.r
\name{step1.4_filter_quality_chimera}
\alias{step1.4_filter_quality_chimera}
\title{Step1.4_filter_quality_chimera}
\usage{
step1.4_filter_quality_chimera(
  folder,
  raw_files,
  fwd_no_unspec,
  rev_no_unspec,
  parameteroptions
)
}
\arguments{
\item{folder}{A string. Location of raw data files.}

\item{raw_files}{A string. List of files from same location.}

\item{fwd_no_unspec}{A character vector. the file name of forward
reads of the input samples where primer sequences were removed.}

\item{rev_no_unspec}{A character vector. Contains the file location of reverse
reads of the input samples where primer sequences were removed.}

\item{parameteroptions}{A dataframe.}
}
\value{
A data frame.
}
\description{
Step1.4_filter_quality_chimera
}
