% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/step2.0_annotate_asvs.r
\name{Step2.0_annotate_ASVs}
\alias{Step2.0_annotate_ASVs}
\title{Title Step2.0_annotate_ASVs}
\usage{
Step2.0_annotate_ASVs(folder, parameteroptions, collated_asv_batches)
}
\arguments{
\item{folder}{A string. Location of raw fastq files}

\item{parameteroptions}{A dataframe,}

\item{collated_asv_batches}{result from Step1..., dataframe}
}
\value{
A dataframe, written to file, combing ASVs across multiple batches
}
\description{
Title Step2.0_annotate_ASVs
}
\examples{
This is the examples section, under Roxygen standard header
}
