% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/step2.1_taxa_allocation.r
\name{Step2.1_taxa_allocation}
\alias{Step2.1_taxa_allocation}
\title{Convert sequence reads to bacteria taxa names using the SILVA v138.1 database
as a reference.}
\usage{
Step2.1_taxa_allocation(folder, collated_asv_batches, taxalevel)
}
\arguments{
\item{folder}{A string. Location of raw data files.}

\item{collated_asv_batches}{A dataframe. Contains the S16 ASV reads, collated across batches.}

\item{taxalevel}{A string. Specifies the taxonomic level used for analysis.}
}
\value{
A dataframe.
}
\description{
Convert sequence reads to bacteria taxa names using the SILVA v138.1 database
as a reference.
}
