% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.0_predict_iqi_multiple.R
\name{step3.0_predict_iqi_multiple}
\alias{step3.0_predict_iqi_multiple}
\title{Generates the predicted IQI values based on the optimized random forest
models.}
\usage{
step3.0_predict_iqi_multiple(folder, AnnotatedASVs, auto_download = FALSE)
}
\arguments{
\item{folder}{A string. Location of raw fastq data files.}

\item{AnnotatedASVs}{A data frame. Contains the bacterial taxa read counts which
should be used to predict the IQI values.}

\item{auto_download}{Logical. If \code{TRUE}, automatically downloads dependencies.}
}
\value{
A data frame. As inputted, with added 'predicted_IQI' column
}
\description{
Generates the predicted IQI values based on the optimized random forest
models.
}
\section{Example usage}{
 myPREDs <- step3.0_predict_iqi(folder = "C:/full/file/path/to/fastqs/", S16_reads = myTAXA)
}

